+++ CHANGELOG.md
@@ -7,10 +7,12 @@ This release adds support to OneView Rest API version 500 for the hardware varia
    - Ethernet Network
    - FC Network
    - FCoE Network
+   - Logical Downlink
    - Logical Interconnect Group
    - Logical Switch Group
    - Network Set
    - SAN Manager
+   - SAS Logical Interconnect
    - SAS Logical Interconnect Group
    - Scopes
    - Switches